requirements.txt
annotated-types==0.7.0
anyio==4.9.0
argcomplete==1.10.3
beautifulsoup4==4.8.2
blis==1.3.0
catalogue==2.0.10
certifi==2025.4.26
cffi==1.17.1
chardet==3.0.4
charset-normalizer==3.4.2
click==8.2.0
cloudpathlib==0.21.0
compressed-rtf==1.0.7
confection==0.1.5
cryptography==44.0.3
cymem==2.0.11
docx2txt==0.9
ebcdic==1.1.1
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
exceptiongroup==1.3.0
extract-msg==0.28.7
fastapi==0.115.12
filelock==3.18.0
fsspec==2025.3.2
h11==0.16.0
hf-xet==1.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.31.1
idna==3.10
IMAPClient==2.1.0
Jinja2==3.1.6
joblib==1.5.0
langcodes==3.5.0
language_data==1.3.0
lxml==5.4.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
murmurhash==1.0.12
networkx==3.4.2
nltk==3.8.1
numpy==2.2.5
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
olefile==0.47
ollama==0.4.8
packaging==25.0
pdf2image==1.17.0
pdfminer.six==20250327
pdfplumber==0.11.6
pillow==11.2.1
preshed==3.0.9
pycparser==2.22
pycryptodome==3.22.0
pydantic==2.11.4
pydantic_core==2.33.2
Pygments==2.19.1
pypdfium2==4.30.1
pytesseract==0.3.13
python-docx==1.1.2
python-dotenv==1.1.0
python-multipart==0.0.20
python-pptx==0.6.23
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==14.0.0
safetensors==0.5.3
shellingham==1.5.4
six==1.12.0
smart-open==7.1.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
spacy==3.8.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SpeechRecognition==3.8.1
srsly==2.5.1
starlette==0.46.2
sympy==1.14.0
thinc==8.3.6
tokenizers==0.21.1
torch==2.7.0
torchvision==0.22.0
tqdm==4.67.1
transformers==4.51.3
triton==3.3.0
typer==0.15.3
typing-inspection==0.4.0
typing_extensions==4.13.2
tzlocal==5.3.1
urllib3==2.4.0
uvicorn==0.34.2
wasabi==1.1.3
weasel==0.4.1
wrapt==1.17.2
xlrd==1.2.0
XlsxWriter==3.2.3
